move druid time
