+++ web-console/src/druid-models/dimension-spec.spec.ts
@@ -20,7 +20,7 @@ import { getDimensionSpecs } from './dimension-spec';
 
 describe('dimension-spec', () => {
   it('getDimensionSpecs', () => {
-    expect(getDimensionSpecs({ header: ['header'], rows: [] }, true)).toMatchInlineSnapshot(`
+    expect(getDimensionSpecs({ header: ['header'], rows: [] }, {}, true)).toMatchInlineSnapshot(`
       Array [
         "header",
       ]
