+++ web-console/src/druid-models/index.ts
@@ -19,11 +19,13 @@
 export * from './compaction-config';
 export * from './compaction-status';
 export * from './lookup-spec';
+export * from './time';
 export * from './timestamp-spec';
 export * from './transform-spec';
 export * from './input-source';
 export * from './input-format';
 export * from './flatten-spec';
+export * from './filter';
 export * from './dimension-spec';
 export * from './metric-spec';
 export * from './ingestion-spec';
