+++ web-console/src/utils/utils.spec.ts
@@ -18,14 +18,9 @@
 
 import { IngestionSpec } from '../druid-models';
 
-import {
-  getColumnTypeFromHeaderAndRows,
-  guessTypeFromSample,
-  updateSchemaWithSample,
-} from './druid-type';
 import { applyCache, headerFromSampleResponse } from './sampler';
 
-describe('test-utils', () => {
+describe('utils', () => {
   const ingestionSpec: IngestionSpec = {
     type: 'index_parallel',
     spec: {
@@ -118,131 +113,6 @@ describe('test-utils', () => {
           "type": "index",
         },
         "type": "index",
-      }
-    `);
-  });
-
-  // it('spec-utils sampleForParser', async () => {
-  //   expect(await sampleForParser(ingestionSpec, 'start', 'abc123')).toMatchInlineSnapshot(
-  //     `Promise {}`,
-  //   );
-  // });
-  //
-  // it('spec-utils SampleSpec', async () => {
-  //   expect(await sampleForConnect(ingestionSpec, 'start')).toMatchInlineSnapshot(`Promise {}`);
-  // });
-  //
-  // it('spec-utils sampleForTimestamp', async () => {
-  //   expect(await sampleForTimestamp(ingestionSpec, 'start', cacheRows)).toMatchInlineSnapshot();
-  // });
-  //
-  // it('spec-utils sampleForTransform', async () => {
-  //   expect(await sampleForTransform(ingestionSpec, 'start', cacheRows)).toMatchInlineSnapshot();
-  // });
-  //
-  // it('spec-utils sampleForFilter', async () => {
-  //   expect(await sampleForFilter(ingestionSpec, 'start', cacheRows)).toMatchInlineSnapshot();
-  // });
-  //
-  // it('spec-utils sampleForSchema', async () => {
-  //   expect(await sampleForSchema(ingestionSpec, 'start', cacheRows)).toMatchInlineSnapshot();
-  // });
-  //
-  // it('spec-utils sampleForExampleManifests', async () => {
-  //   expect(await sampleForExampleManifests('some url')).toMatchInlineSnapshot();
-  // });
-});
-
-describe('druid-type.ts', () => {
-  const ingestionSpec: IngestionSpec = {
-    type: 'index_parallel',
-    spec: {
-      ioConfig: {
-        type: 'index_parallel',
-        inputSource: {
-          type: 'http',
-          uris: ['https://static.imply.io/data/wikipedia.json.gz'],
-        },
-        inputFormat: {
-          type: 'json',
-        },
-      },
-      tuningConfig: {
-        type: 'index_parallel',
-      },
-      dataSchema: {
-        dataSource: 'wikipedia',
-        granularitySpec: {
-          type: 'uniform',
-          segmentGranularity: 'DAY',
-          queryGranularity: 'HOUR',
-        },
-        timestampSpec: {
-          column: 'timestamp',
-          format: 'iso',
-        },
-        dimensionsSpec: {},
-      },
-    },
-  };
-
-  it('spec-utils guessTypeFromSample', () => {
-    expect(guessTypeFromSample([])).toMatchInlineSnapshot(`"string"`);
-  });
-
-  it('spec-utils getColumnTypeFromHeaderAndRows', () => {
-    expect(
-      getColumnTypeFromHeaderAndRows({ header: ['header'], rows: [] }, 'header'),
-    ).toMatchInlineSnapshot(`"string"`);
-  });
-
-  it('spec-utils updateSchemaWithSample', () => {
-    expect(
-      updateSchemaWithSample(ingestionSpec, { header: ['header'], rows: [] }, 'specific', true),
-    ).toMatchInlineSnapshot(`
-      Object {
-        "spec": Object {
-          "dataSchema": Object {
-            "dataSource": "wikipedia",
-            "dimensionsSpec": Object {
-              "dimensions": Array [
-                "header",
-              ],
-            },
-            "granularitySpec": Object {
-              "queryGranularity": "HOUR",
-              "rollup": true,
-              "segmentGranularity": "DAY",
-              "type": "uniform",
-            },
-            "metricsSpec": Array [
-              Object {
-                "name": "count",
-                "type": "count",
-              },
-            ],
-            "timestampSpec": Object {
-              "column": "timestamp",
-              "format": "iso",
-            },
-          },
-          "ioConfig": Object {
-            "inputFormat": Object {
-              "type": "json",
-            },
-            "inputSource": Object {
-              "type": "http",
-              "uris": Array [
-                "https://static.imply.io/data/wikipedia.json.gz",
-              ],
-            },
-            "type": "index_parallel",
-          },
-          "tuningConfig": Object {
-            "type": "index_parallel",
-          },
-        },
-        "type": "index_parallel",
       }
     `);
   });
